requirements.txt
# Core Django and API framework
Django==4.2.16
djangorestframework==3.15.2
drf-spectacular==0.26.4

# Database and caching
psycopg2-binary==2.9.7
redis==4.6.0

# Async processing
celery==5.3.1
django-celery-beat==2.5.0

# HTTP and API clients
requests==2.31.0
httpx==0.24.1

# Security and encryption
cryptography==41.0.3
django-cors-headers==4.2.0

# Configuration and utilities
python-decouple==3.8
xmltodict==0.13.0
django-extensions==3.2.3
dj-database-url==2.1.0

# Production server
gunicorn==21.2.0

# Development and testing
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.3.0

# Monitoring and logging
django-health-check==3.17.0
sentry-sdk==1.32.0

# PDF Generation
reportlab==4.0.7
qrcode[pil]==7.4.2
Pillow==10.1.0

# Excel Generation
openpyxl==3.1.2
